Added PopUps, Loading, Spinner and Gestures

+++ www/js/controllers.js
@@ -227,19 +227,29 @@ angular.module('conFusion.controllers', [])
             
                     }])
 
-.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', function ($scope, menuFactory, favoriteFactory, baseURL) {
+.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicPopup, $ionicLoading, $timeout) {
 
                     $scope.baseURL = baseURL;
                     $scope.shouldShowDelete = false;
+    
+                    $ionicLoading.show({
+                        template: '<ion-spinner></ion-spinner> Loading...'
+                    });
 
                     $scope.favorites = favoriteFactory.getFavorites();
 
                     $scope.dishes = menuFactory.getDishes().query(
                         function (response) {
                             $scope.dishes = response;
+                            $timeout(function () {
+                                $ionicLoading.hide();
+                            }, 1000);                            
                         },
                         function (response) {
                             $scope.message = "Error: " + response.status + " " + response.statusText;
+                            $timeout(function () {
+                                $ionicLoading.hide();
+                            }, 1000);                              
                         });
                     console.log($scope.dishes, $scope.favorites);
 
@@ -250,10 +260,24 @@ angular.module('conFusion.controllers', [])
 
                     $scope.deleteFavorite = function (index) {
 
-                        favoriteFactory.deleteFromFavorites(index);
+                        var confirmPopup = $ionicPopup.confirm({
+                            title: 'Confirm Delete',
+                            template: 'Are you sure you want to delete this item?'
+                        });
+
+                        confirmPopup.then(function (res) {
+                            if (res) {
+                                console.log('Ok to delete');
+                                favoriteFactory.deleteFromFavorites(index);
+                            } else {
+                                console.log('Canceled delete');
+                            }
+                        });
+
                         $scope.shouldShowDelete = false;
 
                     }
+
                 }])
 
 .filter('favoriteFilter', function () {
